chore: move popperWrapper in atoms and tooltip in molecules folder
