+++ core/components/atoms/popperWrapper/index.tsx
@@ -1,0 +1,2 @@
+export { default } from './PopperWrapper';
+export * from './PopperWrapper';
